fix: update unconfigured annotation filters on item change
