+++ src/components/panel/annotations/filters/AnnotationFilters.tsx
@@ -1,4 +1,4 @@
-import { FC } from 'react'
+import { FC, useEffect } from 'react'
 import { usePanel } from '@/contexts/PanelContext.tsx'
 import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover.tsx'
 import { Button } from '@/components/ui/button.tsx'
@@ -6,15 +6,38 @@ import { ListFilter } from 'lucide-react'
 import AnnotationFiltersContent from '@/components/panel/annotations/filters/AnnotationFiltersContent.tsx'
 import { cn } from '@/lib/utils.ts'
 import { SIDEBAR_DEFAULT_WIDTH } from '@/utils/panel.ts'
+import { useConfig } from '@/contexts/ConfigContext.tsx'
 
 
 interface Props {
   className?: string
 }
 const AnnotationFilters: FC<Props> = ({ className }) => {
-  const { usePanelTranslation } = usePanel()
+  const { annotations: annotationsConfig } = useConfig()
+  const {
+    usePanelTranslation,
+    setAnnotationFilters,
+    matchedAnnotationsMap
+  } = usePanel()
+
   const { t } = usePanelTranslation()
 
+  useEffect(() => {
+    // This is for the case where no specific annotation filters were configured.
+    // We extract all occurring types from the annotations that match the text.
+
+    if (annotationsConfig.filters) return
+
+    const uniqueAnnotationTypes: string[] = [
+      ...new Set(Object.keys(matchedAnnotationsMap).map((id) => matchedAnnotationsMap[id].annotation.body['x-content-type']))
+    ]
+
+    setAnnotationFilters({
+      rootSelectionRule: 'multiple',
+      items: uniqueAnnotationTypes.map(type => ({ types: [type], selected: true }))
+    })
+
+  }, [matchedAnnotationsMap])
 
   return <div className={cn('flex flex-col items-center', className)}>
     <Popover>
